+++ docs/en/changelog.rst
@@ -15,6 +15,10 @@ Core
 
 * fixed physics simulation loop freezing, added tests
 
+* fixed "prefix" attribute when colldata models are used.
+
+* added "scalegeometry" attribute to kinbody loading
+
 Inverse Kinematics
 ------------------